ti: bq2597x_charger_munch: Cast void pointer to int without warning
This fixes following warning:
../drivers/power/supply/ti/bq2597x_charger_munch.c:2526:9: warning: cast to smaller integer type 'int' from 'const void *' [-Wvoid-pointer-to-int-cast]
        return (int)of_id->data;
               ^~~~~~~~~~~~~~~~
1 warning generated.

+++ drivers/power/supply/ti/bq2597x_charger_munch.c
@@ -2523,7 +2523,7 @@ static int bq2597x_get_dev_role(struct i2c_client *client)
 
 	dev_info(&client->dev, "%s: matched to %s\n", __func__, of_id->compatible);
 
-	return (int)of_id->data;
+	return (uintptr_t)of_id->data;
 }
 
 static int bq2597x_charger_probe(struct i2c_client *client,
